+++ E-class/e-commerce-backend/src/main/java/com/vn/backend/repository/AttributeRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AttributeRepository extends JpaRepository<Attribute, Integer> {
+public interface AttributeRepository extends JpaRepository<Attribute, Long> {
 
     @EntityGraph(attributePaths = {"attributeValues"})
     List<Attribute> findAll();
